TensorFlow: upstream changes from the afternoon.
Changes:

- futurize --stage2 changes for Python 3 compatibility by @girving.

- Small updates to documentation by @vrv, schuster and others

- Account for failure of std::thread::hardware_concurrency by @ebrevdo.

- More changes for backwards-compatibility tests by Josh

- Updates to python op doc generation by Josh

- Added support for using the best-fit allocator via ConfigProto by @vrv.

- Rename LocalSession to DirectSession, since local was a bad name for
  it.

- Enable tf.nn.moments() to work with tensors of unknown shape by @mrry.
  GITHUB_ISSUE: 139

- Changes for Android build by Andrew.

Base CL: 107645181

+++ tensorflow/python/summary/event_accumulator.py
@@ -1,4 +1,7 @@
 """Takes a generator of values, and accumulates them for a frontend."""
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
 
 import collections
 import threading
@@ -393,7 +396,7 @@ class EventAccumulator(object):
     bucket_limit = list(histo.bucket_limit)
 
     bucket_total = sum(bucket)
-    fraction_weights = [float(10000*x)/bucket_total for x in bucket]
+    fraction_weights = [10000 * x / bucket_total for x in bucket]
     cumsum_weights = _CumulativeSum(fraction_weights)
 
     percentiles = [
